multi: Setup main package and logger
This commit sets up the terminator's main package nested in
the cmd directory. This is done because defers created in
the top level scope of a main method are not executed if
os.Exit() is called.

A logger for a single terminator subsystem is addded and
go modules are initialized.