recommend: surface values and false recommendations
This commit adds a Recommendation stuct which provides the value that
a close recommendation was based on and the close recommendation
boolean. It also adds channels with false recommendations to the
report (rather than only containing the channels recommended for close).

The value is added to provide users more
information about how the recommendation
was reached, and providing the false
recommendations provides better insight
into which channels were and were eligible
for a close recommendation but did not
meet the close criteria.