dataset: return false map when outliers cannot be calcualted
Internally catch error that is thrown when we have too few
values to calcualte outliers and simply return an all false
map rather than exposing the error to calling functions. If
there are too few values to calculate outliers, then no values
are outliers so a false map is appropriate.