recommend: add revenue based recommendations
This commit adds revenue based close recommendations which use the
fees a channel has generated, scaled by the number of confirmations
it has to compare channels. We do not have opening timestamps for
channels at present, so we cannot compare channels over a time range
(because we do not know whether the channel was open or not).

A metric enum is used to identify the datapoint that we wish to use.
In this commit, the rpc server is set to default to uptime calculations
so that the rpc/cli changes can be made in a separate commit.