fiat: add new fiat price estimation package and coincap api query
Add a new package which allows us to obtain historical bitcoin prices.
The first external source we query is coincap. Their api allows
specification of granularity of data - the time period over which
prices are aggregated. The api has a limit on the period of time
you may query per granularity - effectively, a limit on the number
of datapoints they will allow. To allow lower levels of granularity,
we allow breaking up of our call into a maximum of 5 separate api
calls.

This granularity setting also affects the data that is returned when
you are under the limit. If request one day's granularity, then only
query for half a day's data, then it will not return any data points.
To account for this, we add a minimum period check as well.

To account for api failures, we allow 3 retries per api call, with a
sleep inbetween. To prevent blocking, we pass context in.
Note that this PR does not have the actual API call and parsing logic
in it. These will be added in followup commits.