go.mod: update dependencies to lnd v0.10.0-beta.rc5
Multiple PRs that follow this one will require a dependency bump, this
commit separates the dependency bump out to reduce duplication. We can
only update to rc5 due to our dependency on lndclient in the loop repo,
which currently depends on the realease candidate. The release candidate
is sufficient for our purposes, and we can bump further as loop is
updated.