multi: extract pagination logic into separate pagination package
We will require generic pagination in a variety of places. Extract the
existing logic in our revenue package to a generic function that can
be used for any paginated endpoint. Although the existing forwarding
events API uses uint32, we use uint64 because it is used in lnd's
invoice and payments endpoints.