accounting: use last htlc in route to find payments to self
Keysend payments do not have payment requests, or the user may not
have specified the payment request using sendpayment. Instead of using
our payment request to determine whether a payment is made to our own
node, we use the last hop in our htlc route. If the payment has no htlcs
it has not completed yet, so is not relevant to our current accounting
period.

Since we need to query for payments outside of our payment range to
properly detect payments to our own node (eg, a payment made before our
period could have been made to one of our own invoices and settled in
our period), we need to run through all payments to figure out which
are made to ourselves. We allow duplicates provided they are made to
the same location, so that we can account for legacy nodes which have
repeated payment hashes.