mutli: detect swaps for on chain accounting report
Use script matching to identify swap sweeps and timeouts. We add a
separate entry type for these transactions to allow users to separate
out their operational costs (as was done for circular payments). Note
that we do not identify on chain htlcs for swaps (eg, the offered htlc
in loop in), additional information is required to do so.

We could collect all of the transactions that these sweeps spend from
and classify them as on chain swap htlcs, but this would be incomplete
because un-swept htlcs would not be identified yet so we leave this
category out completely rather than offering incomplete information.