bugfix: account for double counting of fee in listchaintxns output
LND includes the fee amount in the total amount for a transaction and
lists it in the fee field. In this change we separate the fee from the
amount so that we do not double count. We do not expect fees for
receives, so we add a check for this.