multi: do not require granularity parameter when fiat is disabled
Previsously, we would try to get the best fiat granularity for a period
even if the disable fiat flag was set. If the period queried is longer
than the longest time that we can query the coincap api for, our request
would fail. This is not practically an issue, because the API allows
20 years of data, but this error does stop setting of a 0 start time,
which is nice for getting an all time report for your node.