multi: add TLS config to RPC server
As a preparation to start the gRPC server using TLS, we make sure the
certificate is created and loaded correctly before starting the server
itself.
We also need to silence the linter that complains about the struct
alignment because we didn't put all boolean variables to the beginning.
But since this struct will only be loaded once per process we care more
about similar options being grouped than memory efficiency.