frdrpc: allow faraday to be used as external subserver
When faraday runs in the same process as lnd (in LiT), it hooks itself
into lnd's RPC server as an external subserver. But because the user
should still be able to use the default faraday macaroon, the faraday
daemon must be able to validate its own macaroons as lnd's macaroon
service doesn't know the root key for it.