fiat: add support for multiple granularities to coingecko API
The coingecko API automatically changes its granularity depending on how
old is the pricing data you are asking for. For 90 days or less, it returns
hourly entries while for any date 91 days or older it returns daily
granularity.

To obtain the most exact results we should always make any reports in that
90 days window. That's why we currently do not support ranges that
include dates older than 90 days.

This commit extends the current API to adapt the range of our queries to
get the most accurate data. That means that if we use this to get the
fiat price of txs that happend between 100 and 1 day ago we will use
daily granularity for the ones older than 90 days and minute granularity
for the ones within the last 90 days.