frdrpc: add POST bindings for more complex calls
Using GET mappings for REST means we have to place every parameter in
the URL query. This is not very convenient if there are many parameters.
But even worse, some types like arrays/slices aren't supported in query
parameters and map types are very awkward to use.
So we add POST mappings for the calls that have many or more complex
request parameters.